+++ lib/pio/open_flow10.rb
@@ -1,5 +1,4 @@
 # Messages
-require 'pio/open_flow10/stats_reply'
 require 'pio/open_flow10/aggregate_stats/reply'
 require 'pio/open_flow10/aggregate_stats/request'
 require 'pio/open_flow10/barrier/reply'
@@ -14,12 +13,14 @@ require 'pio/open_flow10/error/hello_failed'
 require 'pio/open_flow10/features/reply'
 require 'pio/open_flow10/features/request'
 require 'pio/open_flow10/flow_mod'
+require 'pio/open_flow10/flow_removed'
 require 'pio/open_flow10/flow_stats/reply'
 require 'pio/open_flow10/flow_stats/request'
 require 'pio/open_flow10/hello'
 require 'pio/open_flow10/packet_in'
 require 'pio/open_flow10/packet_out'
 require 'pio/open_flow10/port_status'
+require 'pio/open_flow10/stats_reply'
 require 'pio/open_flow10/stats_request'
 
 # Actions